+++ src/Sylius/Bundle/UserBundle/Command/PromoteUserCommand.php
@@ -13,58 +13,18 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\UserBundle\Command;
 
-use Sylius\Component\User\Model\UserInterface;
-use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
+trigger_deprecation(
+    'sylius/user-bundle',
+    '1.13',
+    'The "%s" class is deprecated and will be removed in Sylius 2.0. Use "%s" instead.',
+    PromoteUserCommand::class,
+    \Sylius\Bundle\UserBundle\Console\Command\PromoteUserCommand::class,
+);
 
-class PromoteUserCommand extends AbstractRoleCommand
-{
-    protected static $defaultName = 'sylius:user:promote';
+class_exists(\Sylius\Bundle\UserBundle\Console\Command\PromoteUserCommand::class);
 
-    protected function configure(): void
+if (false) {
+    class PromoteUserCommand
     {
-        $this
-            ->setDescription('Promotes a user by adding roles.')
-            ->setDefinition([
-                new InputArgument('email', InputArgument::REQUIRED, 'Email'),
-                new InputArgument('roles', InputArgument::IS_ARRAY, 'Security roles'),
-                new InputOption('super-admin', null, InputOption::VALUE_NONE, 'Set the user as a super admin'),
-                new InputOption('user-type', null, InputOption::VALUE_REQUIRED, 'User type'),
-            ])
-            ->setHelp(
-                <<<EOT
-The <info>sylius:user:promote</info> command promotes a user by adding security roles
-
-  <info>php app/console sylius:user:promote matthieu@email.com</info>
-EOT
-            )
-        ;
-    }
-
-    protected function executeRoleCommand(InputInterface $input, OutputInterface $output, UserInterface $user, array $securityRoles): void
-    {
-        $error = false;
-        $successMessages = [];
-
-        foreach ($securityRoles as $securityRole) {
-            if ($user->hasRole($securityRole)) {
-                $output->writeln(sprintf('<error>User "%s" already has "%s" security role.</error>', $user->getEmail(), $securityRole));
-                $error = true;
-
-                continue;
-            }
-
-            $user->addRole($securityRole);
-            $successMessages[] = sprintf('Security role <comment>%s</comment> has been added to user <comment>%s</comment>', $securityRole, $user->getEmail());
-        }
-
-        if (!$error) {
-            $output->writeln($successMessages);
-            $this->getEntityManager($input->getOption('user-type'))->flush();
-        } else {
-            $output->writeln(sprintf('<error>No roles added to User "%s".</error>', $user->getEmail()));
-        }
     }
 }
